+++ src/components/header/header.tsx
@@ -1,4 +1,4 @@
-const Header = () => (
+const Header = (): JSX.Element => (
   <header className="header">
     <div className="container">
       <div className="header__wrapper">
