+++ src/components/map/map.tsx
@@ -1,0 +1,3 @@
+const Map = (): JSX.Element => (<section className="cities__map map"></section>);
+
+export default Map;
