--- src/pages/Main/mockData.ts
@@ -1,38 +1,0 @@
-interface ICities {
-  id: number;
-  title: string;
-  link: string;
-}
-
-export const cities: ICities[] = [
-  {
-    id: 1,
-    title: 'Paris',
-    link: '/paris',
-  },
-  {
-    id: 2,
-    title: 'Cologne',
-    link: '/cologne',
-  },
-  {
-    id: 3,
-    title: 'Brussels',
-    link: '/brussels',
-  },
-  {
-    id: 4,
-    title: 'Amsterdam',
-    link: '/amsterdam',
-  },
-  {
-    id: 5,
-    title: 'Hamburg',
-    link: '/hamburg',
-  },
-  {
-    id: 6,
-    title: 'Dusseldorf',
-    link: '/dusseldorf',
-  },
-];
