+++ src/pages/login-page/login-page.tsx
@@ -48,6 +48,7 @@ const LoginPage = (): JSX.Element => (
             </button>
           </form>
         </section>
+
         <section className="locations locations--login locations--current">
           <div className="locations__item">
             <a className="locations__item-link" href="#">
