+++ src/pages/main-page/components/nav-tabs/nav-tabs.tsx
@@ -1,0 +1,19 @@
+import { CITIES } from '../../../../const';
+
+const NavTabs = (): JSX.Element => (
+  <div className="tabs">
+    <section className="locations container">
+      <ul className="locations__list tabs__list">
+        {CITIES.map((city) => (
+          <li className="locations__item" key={city}>
+            <a className="locations__item-link tabs__item" href="#">
+              <span>{city}</span>
+            </a>
+          </li>
+        ))}
+      </ul>
+    </section>
+  </div>
+);
+
+export default NavTabs;
