+++ src/pages/main-page/components/sort-places/sort-places.tsx
@@ -1,0 +1,37 @@
+interface IAppScreenProps {
+  offer: number;
+}
+
+const sortTypes: string[] = [
+  'Popular',
+  'Price: low to high',
+  'Price: high to low',
+  'Top rated first',
+];
+
+const SortPlaces = ({ offer }: IAppScreenProps): JSX.Element => (
+  <div>
+    <h2 className="visually-hidden">Places</h2>
+    <b className="places__found">{offer} places to stay in Amsterdam</b>
+
+    <form className="places__sorting" action="#" method="get">
+      <span className="places__sorting-caption">Sort by</span>
+      <span className="places__sorting-type">
+        Popular
+        <svg className="places__sorting-arrow" width="7" height="4">
+          <use xlinkHref="#icon-arrow-select"></use>
+        </svg>
+      </span>
+
+      <ul className="places__options places__options--custom places__options--opened">
+        {sortTypes.map((type) => (
+          <li className="places__option" key={type}>
+            {type}
+          </li>
+        ))}
+      </ul>
+    </form>
+  </div>
+);
+
+export default SortPlaces;
